# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Faren API - personalized profile platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [social]
      summary: Follow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfollowUser
      tags: [social]
      summary: Unfollow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{username}/like:
    post:
      operationId: likeProfile
      tags: [social]
      summary: Like a profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /analytics/record-view:
    post:
      operationId: recordProfileView
      tags: [analytics]
      summary: Record a profile view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordViewBody"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /discover/trending:
    get:
      operationId: getTrendingProfiles
      tags: [social]
      summary: Get trending profiles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Trending profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicProfile"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    PublicProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        backgroundOpacity:
          type: ["number", "null"]
        glowColor:
          type: ["string", "null"]
        cursorStyle:
          type: ["string", "null"]
        musicUrl:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLink"
        discordConnected:
          type: boolean
        discordUsername:
          type: ["string", "null"]
        discordAvatarUrl:
          type: ["string", "null"]
        discordStatus:
          type: ["string", "null"]
        discordActivity:
          type: ["string", "null"]
        discordStatusEmoji:
          type: ["string", "null"]
        musicConnected:
          type: boolean
        musicService:
          type: ["string", "null"]
        nowPlaying:
          $ref: "#/components/schemas/NowPlaying"
        followersCount:
          type: integer
        followingCount:
          type: integer
        likesCount:
          type: integer
        viewsCount:
          type: integer
        isFollowing:
          type: boolean
        hasLiked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - badges
        - links
        - discordConnected
        - musicConnected
        - followersCount
        - followingCount
        - likesCount
        - viewsCount
        - isFollowing
        - hasLiked
        - createdAt
    ProfileLink:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        label:
          type: string
        url:
          type: string
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - platform
        - label
        - url
        - sortOrder
...
    NowPlaying:
      type: object
      properties:
        isPlaying:
          type: boolean
        title:
          type: ["string", "null"]
        artist:
          type: ["string", "null"]
        album:
          type: ["string", "null"]
        albumArt:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        externalUrl:
          type: ["string", "null"]
        progress:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
      required:
        - isPlaying
...
    RecordViewBody:
      type: object
      properties:
        username:
          type: string
        country:
          type: ["string", "null"]
        device:
          type: ["string", "null"]
      required:
        - username